+++ itou/www/apply/forms.py
@@ -20,43 +20,12 @@ from itou.eligibility.models import AdministrativeCriteria
 from itou.files.forms import ItouFileField
 from itou.job_applications import enums as job_applications_enums
 from itou.job_applications.models import JobApplication, PriorAction
-from itou.users.forms import JobSeekerProfileFieldsMixin
-from itou.users.models import JobSeekerProfile, User
+from itou.users.models import JobSeekerProfile
 from itou.utils import constants as global_constants
 from itou.utils.templatetags.str_filters import mask_unless
 from itou.utils.types import InclusiveDateRange
 from itou.utils.widgets import DuetDatePickerWidget
 from itou.www.companies_views.forms import JobAppellationAndLocationMixin
-
-
-class CheckJobSeekerInfoForm(JobSeekerProfileFieldsMixin, forms.ModelForm):
-    PROFILE_FIELDS = ["birthdate", "pole_emploi_id", "lack_of_pole_emploi_id_reason"]
-
-    class Meta:
-        model = User
-        fields = [
-            "phone",
-        ]
-        help_texts = {
-            "birthdate": "Au format JJ/MM/AAAA, par exemple 20/12/1978.",
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["birthdate"].required = True
-        self.fields["birthdate"].widget = DuetDatePickerWidget(
-            {
-                "min": DuetDatePickerWidget.min_birthdate(),
-                "max": DuetDatePickerWidget.max_birthdate(),
-            }
-        )
-
-    def clean(self):
-        super().clean()
-        JobSeekerProfile.clean_pole_emploi_fields(self.cleaned_data)
-        JobSeekerProfile.clean_nir_title_birthdate_fields(
-            self.cleaned_data | {"nir": self.instance.jobseeker_profile.nir}, remind_nir_in_error=True
-        )
 
 
 class ApplicationJobsForm(forms.ModelForm):
