+++ itou/www/job_seekers_views/views.py
@@ -26,9 +26,10 @@ from itou.utils.emails import redact_email_address
 from itou.utils.pagination import ItouPaginator
 from itou.utils.session import SessionNamespace, SessionNamespaceRequiredMixin
 from itou.utils.urls import get_safe_url
-from itou.www.apply.views.submit_views import ApplyStepBaseView, ApplyStepForSenderBaseView
+from itou.www.apply.views.submit_views import ApplicationBaseView, ApplyStepBaseView, ApplyStepForSenderBaseView
 
 from .forms import (
+    CheckJobSeekerInfoForm,
     CheckJobSeekerNirForm,
     CreateOrUpdateJobSeekerStep1Form,
     CreateOrUpdateJobSeekerStep2Form,
@@ -206,7 +207,7 @@ class CheckNIRForJobSeekerView(ApplyStepBaseView):
         if self.job_seeker.jobseeker_profile.nir:
             return HttpResponseRedirect(
                 reverse(
-                    "apply:step_check_job_seeker_info",
+                    "job_seekers_views:check_job_seeker_info",
                     kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
                 )
             )
@@ -215,7 +216,7 @@ class CheckNIRForJobSeekerView(ApplyStepBaseView):
 
     def post(self, request, *args, **kwargs):
         next_url = reverse(
-            "apply:step_check_job_seeker_info",
+            "job_seekers_views:check_job_seeker_info",
             kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
         )
         if self.form.is_valid():
@@ -740,7 +741,7 @@ class UpdateJobSeekerStep1View(UpdateJobSeekerBaseView):
     template_name = "job_seekers_views/create_or_update_job_seeker/step_1.html"
 
     previous_apply_url = "apply:application_jobs"
-    previous_hire_url = "apply:check_job_seeker_info_for_hire"
+    previous_hire_url = "job_seekers_views:check_job_seeker_info_for_hire"
     next_apply_url = "job_seekers_views:update_job_seeker_step_2"
     next_hire_url = "job_seekers_views:update_job_seeker_step_2_for_hire"
 
@@ -905,7 +906,7 @@ class UpdateJobSeekerStepEndView(UpdateJobSeekerBaseView):
     previous_apply_url = "job_seekers_views:update_job_seeker_step_3"
     previous_hire_url = "job_seekers_views:update_job_seeker_step_3_for_hire"
     next_apply_url = "apply:application_jobs"
-    next_hire_url = "apply:check_job_seeker_info_for_hire"
+    next_hire_url = "job_seekers_views:check_job_seeker_info_for_hire"
 
     def __init__(self):
         super().__init__()
@@ -989,3 +990,72 @@ class UpdateJobSeekerStepEndView(UpdateJobSeekerBaseView):
 
     def get_context_data(self, **kwargs):
         return super().get_context_data(**kwargs) | {"profile": self.profile, "progress": "80"}
+
+
+class CheckJobSeekerInformations(ApplicationBaseView):
+    """
+    Ensure the job seeker has all required info.
+    """
+
+    template_name = "job_seekers_views/check_job_seeker_info.html"
+
+    def __init__(self):
+        super().__init__()
+
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+
+        self.form = CheckJobSeekerInfoForm(instance=self.job_seeker, data=request.POST or None)
+
+    def get(self, request, *args, **kwargs):
+        # Check required info that will allow us to find a pre-existing approval.
+        has_required_info = self.job_seeker.jobseeker_profile.birthdate and (
+            self.job_seeker.jobseeker_profile.pole_emploi_id
+            or self.job_seeker.jobseeker_profile.lack_of_pole_emploi_id_reason
+        )
+        if has_required_info:
+            return HttpResponseRedirect(
+                reverse(
+                    "apply:step_check_prev_applications",
+                    kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+                )
+            )
+
+        return super().get(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        if self.form.is_valid():
+            self.form.save()
+            return HttpResponseRedirect(
+                reverse(
+                    "apply:step_check_prev_applications",
+                    kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+                )
+            )
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+        }
+
+
+class CheckJobSeekerInformationsForHire(ApplicationBaseView):
+    """
+    Ensure the job seeker has all required info.
+    """
+
+    template_name = "job_seekers_views/check_job_seeker_info_for_hire.html"
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        assert self.hire_process
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "profile": self.job_seeker.jobseeker_profile,
+            "back_url": reverse("job_seekers_views:check_nir_for_hire", kwargs={"company_pk": self.company.pk}),
+        }
